requirements.txt
torch>=2.2.0
torch-geometric>=2.5.0
pandas>=2.0.0
polars>=1.8.0
pyarrow>=15.0.0
numpy>=1.24.0
pot>=0.9.3 # Python Optimal Transport
geomloss>=0.2.6 # For Sinkhorn divergences
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.2
requests>=2.31.0
tqdm>=4.65.0
mlflow>=2.10.0 # Standard choice for experiment tracking
pytest>=8.0.0
jupyter>=1.0.0 # for exploratory analysis
python-dotenv>=1.0.0 # for .env files
entsoe-py>=0.6.2
